#268 rename: Change UserControls.tsx location
